+++ src/core/services/Provider.java
@@ -1,5 +1,6 @@
 package core.services;
 
+import core.interfaces.IProvider;
 import estorePojo.exceptions.UnknownItemException;
 
 import java.util.HashMap;
@@ -28,7 +29,7 @@ public class Provider implements IProvider {
         if (!itemPrices.containsKey(item))
             throw new UnknownItemException("Item " + item + " is not an item delivered by this provider.");
 
-        Double price = (Double) itemPrices.get(item);
+        Double price = itemPrices.get(item);
         return price.doubleValue();
     }
